fix: size killers based on max ply (#346)
Don't size killers table by `MAX_DEPTH`, use `MAX_PLY` instead. 

```
Elo   | 0.62 +- 2.89 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.94 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 20108 W: 5613 L: 5577 D: 8918
Penta | [365, 2156, 4971, 2202, 360]
```
https://openbench.nocturn9x.space/test/6850/

bench: 1352109

+++ engine/src/move_picker.rs
@@ -100,7 +100,7 @@ impl MovePicker {
     /// Creates a `MovePicker` for the main negamax search.
     ///
     /// No moves are generated upfront; generation is deferred to the stage machine.
-    pub(crate) fn new(tt_move: Option<Move>, killers_table: &KillerMovesTable, ply: u8) -> Self {
+    pub(crate) fn new(tt_move: Option<Move>, killers_table: &KillerMovesTable, ply: usize) -> Self {
         let killer_slice = killers_table.get(ply);
         let killers = [
             killer_slice.first().copied().unwrap_or(None),
